terraform update
