+++ terraform/aws_lambda_functions/clipboard_disconnect/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "on-disconnect",
+  "version": "1.0.0",
+  "description": "ondisconnect example for WebSockets on API Gateway",
+  "main": "src/app.js",
+  "author": "SAM CLI",
+  "license": "MIT",
+  "dependencies": {},
+  "devDependencies": {
+    "aws-sdk": "^2.739.0"
+  }
+}
