+++ terraform/aws_lambda_functions/clipboard_sendmessage/package.json
@@ -1,0 +1,13 @@
+{
+    "name": "send-message",
+    "version": "1.0.0",
+    "description": "sendmessage example for WebSockets on API Gateway",
+    "main": "src/app.js",
+    "author": "SAM CLI",
+    "license": "MIT",
+    "dependencies": {},
+    "devDependencies": {
+      "aws-sdk": "^2.739.0"
+    }
+  }
+
