--- terraform/lambda_clipboard_onconnect/package.json
@@ -1,13 +1,0 @@
-{
-    "name": "on-connect",
-    "version": "1.0.0",
-    "description": "onconnect example for WebSockets on API Gateway",
-    "main": "index.js",
-    "author": "SAM CLI",
-    "license": "MIT",
-    "dependencies": {},
-    "devDependencies": {
-      "aws-sdk": "^2.739.0"
-    }
-}
-
