--- terraform/lambda_clipboard_sendmessage/index.js
@@ -1,46 +1,0 @@
-const AWS = require('aws-sdk');
-
-const ddb = new AWS.DynamoDB.DocumentClient({ apiVersion: '2012-08-10', region: process.env.AWS_REGION });
-
-const { TABLE_NAME } = process.env;
-
-exports.lambda_handler = async event => {
-  let connectionData;
-  
-  try {
-    connectionData = await ddb.scan({ TableName: TABLE_NAME, ProjectionExpression: 'connectionId' }).promise();
-  } catch (e) {
-    console.log("MY SENDMESSAGE ERROR",err)
-    return { statusCode: 500, body: e.stack };
-  }
-  
-  const apigwManagementApi = new AWS.ApiGatewayManagementApi({
-    apiVersion: '2018-11-29',
-    endpoint: event.requestContext.domainName + '/' + event.requestContext.stage
-  });
-  
-  const postData = JSON.parse(event.body).data;
-  
-  const postCalls = connectionData.Items.map(async ({ connectionId }) => {
-    try {
-      await apigwManagementApi.postToConnection({ ConnectionId: connectionId, Data: postData }).promise();
-    } catch (e) {
-      if (e.statusCode === 410) {
-        console.log(`Found stale connection, deleting ${connectionId}`);
-        await ddb.delete({ TableName: TABLE_NAME, Key: { connectionId } }).promise();
-      } else {
-        console.log("MY SENDMESSAGE ERROR",e)
-        throw e;
-      }
-    }
-  });
-  
-  try {
-    await Promise.all(postCalls);
-  } catch (e) {
-    console.log("MY SENDMESSAGE ERROR",e)
-    return { statusCode: 500, body: e.stack };
-  }
-
-  return { statusCode: 200, body: 'Data sent.' };
-};
