+++ terraform/output.tf
@@ -1,11 +1,11 @@
 
-# output "url" {
-#   value = "https://s3.${var.region}.amazonaws.com/${aws_s3_bucket.account.id}/index.html"
-# }
+output "url" {
+  value = "https://s3.${var.region}.amazonaws.com/${aws_s3_bucket.account.id}/index.html"
+}
 
-# output "websocket_api" {
-#   value = aws_apigatewayv2_stage.stage_prod.invoke_url
-# }
+output "websocket_api" {
+  value = aws_apigatewayv2_stage.stage_prod.invoke_url
+}
 
 output "aws_cognito_user_pool_id" {
   value = aws_cognito_user_pool.pool.id
