+++ terraform/templates/awsconfig.json
