+++ terraform/terraform.auto.tfvars
@@ -1,0 +1,4 @@
+
+# Application Definition 
+app_name        = "s3fileshare" # Do NOT enter any spaces
+app_environment = "dev"       # Dev, Test, Staging, Prod, etc
